Refactor BearerAuth into connections.py and remove duplicate lines by setting the BearerAuth on the session rather than per-request
